7A_custom_airflow_oprators/requirements.txt
apache-airflow-providers-http
pandas
llama-index